\name{dat1}
\alias{dat1}
\docType{data}
\title{
Public data risk no. 1
}
\description{
Dataset named Public data risk no. 1 consisting in 1000 of simulated data.
}
\usage{data(dat1)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{y}}{a numeric vector}
  }
}


\references{
Bolance, C.; Guillen, M.; Gustafsson, J. & Nielsen, J. P. Quantitative Operational Risk Models Chapman & Hall/CRC, 2012
}
\examples{
data(dat1)
}
\keyword{datasets}
