\name{dat2}
\alias{dat2}
\docType{data}
\title{
Public data risk no. 2
}
\description{
Dataset named Public data risk no. 1 consisting in 400 of simulated data.
}
\usage{data(dat2)}
\format{
  A data frame with 400 observations on the following variable.
  \describe{
    \item{\code{y}}{a numeric vector}
  }
}

\references{
Bolance, C.; Guillen, M.; Gustafsson, J. & Nielsen, J. P. Quantitative Operational Risk Models Chapman & Hall/CRC, 2012
}
\examples{
data(dat2)
}
\keyword{datasets}
