\name{OCA-package}
\alias{OCA-package}
\alias{OCA}
\docType{package}
\title{
Optimal Capital Allocation Principles
}
\description{
\code{OCA} computes optimal capital allocation based on some standard principles such as Haircut, Overbeck type II and the Covariance Allocation Principle. Also it provides some shortcuts for obtaining the Value at Risk and the Expectation Shortfall, using both the normal and the t-student distribution.
}
\details{
 \tabular{ll}{
 Package: \tab OCA\cr
 Type:    \tab Package\cr
 Version: \tab 0.3\cr
 Date:    \tab 2021-01-27\cr
 License: \tab GPL-2\cr
 }
}

\author{
Jilber Urbina

Maintainer: Jilber Urbina <jurbina@cipadla.com>

}
\references{
Dhaene J., Tsanakas A., Valdez E. and Vanduffel S. (2011). \emph{Optimal capital allocation principles}. The Journal of Risk and Insurance. 79(1):1-28.

McNeil, A. J.; Frey, R. and Embrechts, P. \emph{Quantitative risk management: concepts, techniques and tools}. Princeton University Press, 2005

Urbina, J. (2013) \emph{Quantifying Optimal Capital Allocation Principles based on Risk Measures.} Master Thesis, Universitat Politècnica de Catalunya. http://hdl.handle.net/2099.1/19443.

Urbina, J. and Guillén, M. (2014). \emph{An application of capital allocation principles to operational risk and the cost of fraud}. Expert Systems with Applications. 41(16):7023-7031.


}

\keyword{ package }
